03/02/2021

+++ models/unit.go
@@ -6,14 +6,14 @@ import (
 
 // Unit of list
 type Unit struct {
-	DonviID int
+	DonviID int64
 	TenDV   string
 }
 
 // MarshalJSON return the json object of Unit
 func (u *Unit) MarshalJSON() ([]byte, error) {
 	return json.Marshal(&struct {
-		DonviID int    `json:"donvi_id"`
+		DonviID int64  `json:"donvi_id"`
 		TenDV   string `json:"ten_dv"`
 	}{
 		DonviID: u.DonviID,
